Reorganize storage attendance into package
