+++ src/main/java/seedu/address/storage/JsonAdaptedStudent.java
@@ -22,7 +22,7 @@ import seedu.address.model.student.Telegram;
 /**
  * Jackson-friendly version of {@link Student}.
  */
-class JsonAdaptedStudent {
+public class JsonAdaptedStudent {
 
     public static final String MISSING_FIELD_MESSAGE_FORMAT = "Student's %s field is missing!";
