+++ src/main/java/seedu/address/storage/JsonAdaptedTag.java
@@ -9,7 +9,7 @@ import seedu.address.model.components.tag.Tag;
 /**
  * Jackson-friendly version of {@link Tag}.
  */
-class JsonAdaptedTag {
+public class JsonAdaptedTag {
 
     private final String tagName;
