+++ src/main/java/seedu/address/storage/JsonAdaptedUuid.java
@@ -10,7 +10,7 @@ import seedu.address.commons.exceptions.IllegalValueException;
 /**
  * Jackson-friendly version of {@link java.util.UUID}.
  */
-class JsonAdaptedUuid {
+public class JsonAdaptedUuid {
 
     public static final String MESSAGE_NULL_UUID = "UUID is null";
     public static final String MESSAGE_INVALID_UUID = "UUID is invalid";
